+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,55 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.domain.User;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+
+    public TxUserService(UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(long id, String newPassword, String createBy) {
+        DataSource dataSource = DataSourceConfig.getInstance();
+        Connection conn = DataSourceUtils.getConnection(dataSource);
+
+        try {
+            conn.setAutoCommit(false);
+            userService.changePassword(id, newPassword, createBy);
+            conn.commit();
+        } catch (SQLException e) {
+            rollback(conn);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(conn, dataSource);
+        }
+    }
+
+    private void rollback(Connection conn) {
+        try {
+            if (conn != null) {
+                conn.rollback();
+            }
+        } catch (SQLException ignored) {
+        }
+    }
+}
