+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,64 +1,12 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
+public interface UserService {
 
-public class UserService {
+    User findById(long id);
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    void save(User user);
 
-    public UserService(UserDao userDao, UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
-
-    public User findById(long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(long id, String newPassword, String createBy) {
-        DataSource dataSource = DataSourceConfig.getInstance();
-        Connection conn = DataSourceUtils.getConnection(dataSource);
-
-        try {
-            conn.setAutoCommit(false);
-            change(id, newPassword, createBy);
-            conn.commit();
-        } catch (SQLException e) {
-            rollback(conn);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(conn, dataSource);
-        }
-    }
-
-    private void change(long id, String newPassword, String createBy) {
-        User user = findById(id);
-        user.changePassword(newPassword);
-        userDao.update(user);
-        userHistoryDao.log(new UserHistory(user, createBy));
-    }
-
-    private void rollback(Connection conn) {
-        try {
-            if (conn != null) {
-                conn.rollback();
-            }
-        } catch (SQLException ignored) {
-        }
-    }
+    void changePassword(long id, String newPassword, String createdBy);
 }
